requirements.txt
transformers==4.35.0
sentence-transformers==2.2.2
datasets==2.14.0
scikit-learn==1.3.0
rank_bm25==0.2.2
hazm==0.7.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
tqdm==4.65.0
torch>=1.13.0